fix(auto): remap shared docker-git root for compose
