+++ packages/lib/src/shell/docker.ts
@@ -5,71 +5,12 @@ import type { PlatformError } from "@effect/platform/Error"
 import { Duration, Effect, pipe, Schedule } from "effect"
 
 import { runCommandCapture, runCommandExitCode, runCommandWithExitCodes } from "./command-runner.js"
-import { resolveDockerVolumeHostPath } from "./docker-auth.js"
+import { composeSpec, resolveDockerComposeEnv } from "./docker-compose-env.js"
+import { parseInspectNetworkEntry } from "./docker-inspect-parse.js"
 import { CommandFailedError, DockerCommandError } from "./errors.js"
 
 export { classifyDockerAccessIssue, ensureDockerDaemonAccess } from "./docker-daemon-access.js"
 export { parseDockerPublishedHostPorts, runDockerPsPublishedHostPorts } from "./docker-published-ports.js"
-
-const composeSpec = (cwd: string, args: ReadonlyArray<string>) => ({
-  cwd,
-  command: "docker",
-  args: ["compose", "--ansi", "never", "--progress", "plain", ...args]
-})
-
-const resolveEnvValue = (key: string): string | null => {
-  const value = process.env[key]?.trim()
-  return value && value.length > 0 ? value : null
-}
-
-const trimTrailingSlash = (value: string): string => {
-  let end = value.length
-  while (end > 0) {
-    const char = value[end - 1]
-    if (char !== "/" && char !== "\\") {
-      break
-    }
-    end -= 1
-  }
-  return value.slice(0, end)
-}
-
-const resolveProjectsRootCandidate = (): string | null => {
-  const explicit = resolveEnvValue("DOCKER_GIT_PROJECTS_ROOT")
-  if (explicit !== null) {
-    return explicit
-  }
-
-  const home = resolveEnvValue("HOME") ?? resolveEnvValue("USERPROFILE")
-  return home === null ? null : `${trimTrailingSlash(home)}/.docker-git`
-}
-
-const resolveComposeEnv = (
-  cwd: string
-): Effect.Effect<Readonly<Record<string, string>>, never, CommandExecutor.CommandExecutor> =>
-  Effect.gen(function*(_) {
-    const projectsRoot = resolveProjectsRootCandidate()
-    if (projectsRoot === null) {
-      return {}
-    }
-
-    const remappedProjectsRoot = yield* _(resolveDockerVolumeHostPath(cwd, projectsRoot))
-    return remappedProjectsRoot === projectsRoot ? {} : { DOCKER_GIT_PROJECTS_ROOT_HOST: remappedProjectsRoot }
-  })
-
-const parseInspectNetworkEntry = (line: string): ReadonlyArray<readonly [string, string]> => {
-  const idx = line.indexOf("=")
-  if (idx <= 0) {
-    return []
-  }
-  const network = line.slice(0, idx).trim()
-  const ip = line.slice(idx + 1).trim()
-  if (network.length === 0 || ip.length === 0) {
-    return []
-  }
-  const entry: readonly [string, string] = [network, ip]
-  return [entry]
-}
 
 const runCompose = (
   cwd: string,
@@ -77,7 +18,7 @@ const runCompose = (
   okExitCodes: ReadonlyArray<number>
 ): Effect.Effect<void, DockerCommandError | PlatformError, CommandExecutor.CommandExecutor> =>
   Effect.gen(function*(_) {
-    const env = yield* _(resolveComposeEnv(cwd))
+    const env = yield* _(resolveDockerComposeEnv(cwd))
     yield* _(
       runCommandWithExitCodes(
         {
@@ -96,7 +37,7 @@ const runComposeCapture = (
   okExitCodes: ReadonlyArray<number>
 ): Effect.Effect<string, DockerCommandError | PlatformError, CommandExecutor.CommandExecutor> =>
   Effect.gen(function*(_) {
-    const env = yield* _(resolveComposeEnv(cwd))
+    const env = yield* _(resolveDockerComposeEnv(cwd))
     return yield* _(
       runCommandCapture(
         {
